requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers==4.56.2
accelerate>=0.30.0
bitsandbytes>=0.43.0
datasets>=2.18.0
pillow>=10.0.0
scikit-learn>=1.2.0
matplotlib>=3.8.0
tqdm>=4.66.0
numpy>=1.24.0,<2
fsspec<=2026.2.0
qwen-vl-utils>=0.0.14
